+++ srcs/number/ft_itoa.c
@@ -10,9 +10,9 @@ char* ft_itoa(int n) {
     if (nbr == NULL)
         return NULL;
     nbr[len] = '\0';
-    unsigned int nb = n;
+    unsigned int nb = (unsigned int)n;
     if (n < 0)
-        nb = -n;
+        nb = (unsigned int)-n;
     while (len--) {
         nbr[len] = "0123456789"[nb % 10];
         nb /= 10;
